More file and function rearranging
