+++ python/cusignal/utils/_compile_kernels.py
@@ -20,25 +20,25 @@ from enum import Enum
 
 from ._caches import _cupy_kernel_cache, _numba_kernel_cache
 
-from ..filtering._lfilter_kernels import _cupy_lfilter_src
-from ..convolution._convolve_kernels import (
+from ..filtering._lfilter_cuda import _cupy_lfilter_src
+from ..convolution._convolution_cuda import (
     _cupy_convolve_src,
     _cupy_convolve_2d_src,
     _numba_convolve_2d,
     _numba_convolve_2d_signature,
 )
-from ..convolution._correlate_kernels import (
+from ..convolution._convolution_cuda import (
     _cupy_correlate_src,
     _cupy_correlate_2d_src,
     _numba_correlate_2d,
 )
-from ..spectral_analysis._spectral_kernels import (
+from ..spectral_analysis._spectral_cuda import (
     _cupy_lombscargle_src,
     _numba_lombscargle,
     _numba_lombscargle_signature,
 )
 
-from ..filtering._upfirdn_kernels import (
+from ..filtering._upfirdn_cuda import (
     _cupy_upfirdn_1d_src,
     _cupy_upfirdn_2d_src,
     _numba_upfirdn_1d,
